requirements.txt
attrs==25.4.0
Automat==25.4.16
constantly==23.10.4
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
mss==10.1.0
numpy==2.3.4
pillow==12.0.0
pynput==1.8.1
six==1.17.0
Twisted==25.5.0
typing_extensions==4.15.0
zope.interface==8.0.1
